Resolve conflicta after 6a75496 [Driver] Define LinkOption and fix forwarded options to GCC for linking
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/262265
Tested-by: Dmitriev, Valery N <valery.n.dmitriev@intel.com>

+++ llvm/include/llvm/Option/OptTable.h
@@ -48,7 +48,7 @@ public:
     unsigned ID;
     unsigned char Kind;
     unsigned char Param;
-    unsigned short Flags;
+    unsigned Flags; // INTEL
     unsigned short GroupID;
     unsigned short AliasID;
     const char *AliasArgs;
